Move `teleport/DesktopSession` to `shared/components/DesktopSession`
