+++ web/packages/shared/components/DesktopSession/TopBar.tsx
