+++ web/packages/shared/components/DesktopSession/index.ts
