+++ web/packages/teleport/src/lib/util.ts
@@ -29,18 +29,6 @@ export const openNewTab = (url: string) => {
   document.body.removeChild(element);
 };
 
-// Adapted from https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/digest#converting_a_digest_to_a_hex_string
-export async function Sha256Digest(
-  message: string,
-  encoder: TextEncoder = new TextEncoder()
-): Promise<string> {
-  const msgUint8 = encoder.encode(message); // encode as (utf-8) Uint8Array
-  const hashBuffer = await crypto.subtle.digest('SHA-256', msgUint8); // hash the message
-  const hashArray = Array.from(new Uint8Array(hashBuffer)); // convert buffer to byte array
-  const hashHex = hashArray.map(b => b.toString(16).padStart(2, '0')).join(''); // convert bytes to hex string
-  return hashHex;
-}
-
 export type TshLoginCommand = {
   authType: AuthType;
   clusterId?: string;
